openvino/requirements.txt
fastapi==0.101.1
numpy==1.21.6
opencv_contrib_python==4.6.0.66
opencv_python_headless==4.6.0.66
openvino==2023.0.1
paddlepaddle==2.5.1
pyclipper==1.3.0.post4
python-dotenv==1.0.0
Shapely==2.0.1
uvicorn==0.23.2
python-multipart==0.0.6
psutil==5.9.5
